+++ tests/testthat/test-get_metadata.R
@@ -2,7 +2,12 @@ context("test-get_metadata.R")
 
 test_that("get_metadata works", {
   library(xml2)
-  data_url <- Sys.getenv("ODKC_TEST_URL")
-  md <- get_metadata(data_url)
-  expect_equal(class(md), "list")
+  md <- get_metadata(
+    Sys.getenv("ODKC_TEST_PID"),
+    Sys.getenv("ODKC_TEST_FID"),
+    url = Sys.getenv("ODKC_TEST_URL"),
+    un = Sys.getenv("ODKC_TEST_UN"),
+    pw = Sys.getenv("ODKC_TEST_PW")
+  )
+  testthat::expect_equal(class(md), "list")
 })